Move OutboundLink to its own file.
