+++ frontend/lib/ui/customer-support-link.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { OutboundLink } from "../analytics/google-analytics";
+import { OutboundLink } from "./outbound-link";
 
 const CUSTOMER_SUPPORT_EMAIL = "support@justfix.nyc";
